requirements.txt
# Install coincurve first to avoid build issues with python 3.14
# Building from scratch requires pkg-config and ninja locally
coincurve==20.0.0
# Rest
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==26.1.0
base58==2.1.1
base58check==1.0.2
bitarray==3.8.1
bitcoin-utils==0.8.1
blockfrost-python==0.7.0
cachetools==7.0.5
cardano-tools==2.1.0
cbor2==5.9.0
cbor2pure==5.8.0
certifi==2026.2.25
certvalidator==0.11.1
cffi==2.0.0
charset-normalizer==3.4.6
ckzg==2.1.7
coloredlogs==15.0.1
cose==0.9.dev8
crcmod==1.7
cryptography==46.0.5
cytoolz==1.1.0
docker==7.1.0
ecdsa==0.19.1
ECPy==1.2.5
ed25519-blake2b==1.4.1
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==6.0.0b1
frozendict==2.4.7
frozenlist==1.8.0
hdwallet==3.6.1
hexbytes==1.3.1
humanfriendly==10.0
idna==3.11
mnemonic==0.21
mpmath==1.3.0
multidict==6.7.1
ogmios==1.4.3
orjson==3.11.7
oscrypto==1.3.0
parsimonious==0.10.0
pexpect==4.9.0
pprintpp==0.4.0
propcache==0.4.1
ptyprocess==0.7.0
pycardano==0.19.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyNaCl==1.6.2
python-dotenv==1.2.2
pyunormalize==17.0.0
regex==2026.4.4
requests==2.32.5
rlp==4.1.0
setuptools==82.0.1
six==1.17.0
sympy==1.14.0
toolz==1.1.0
typeguard==4.5.1
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
web3==7.15.0
websocket-client==1.9.0
websockets==15.0.1
yarl==1.23.0